requirements.txt
numpy==1.26.1
scikit-learn==1.3.2
matplotlib==3.8.1
boto3==1.28.81
opencv-python==4.8.1.78
python-dotenv==1.0.0
openai==1.1.2
flask==3.0.0
requests==2.31.0